+++ core/tabs/applications-setup/design-tools/blender.sh
@@ -17,7 +17,9 @@ installBlender() {
 				fi
 	            ;;
 	        *)
-	            "$ESCALATION_TOOL" flatpak install --noninteractive org.blender.Blender
+	        	if command_exists flatpak; then
+	            	"$ESCALATION_TOOL" flatpak install --noninteractive org.blender.Blender
+	            fi
 	            exit 1
 	            ;;
 	    esac
