+++ core/tabs/applications-setup/design-tools/mypaint.sh
@@ -17,7 +17,9 @@ installMyPaint() {
 				fi
 	            ;;
 	        *)
-	            "$ESCALATION_TOOL" flatpak install --noninteractive org.mypaint.MyPaint
+	        	if command_exists flatpak; then
+	            	"$ESCALATION_TOOL" flatpak install --noninteractive org.mypaint.MyPaint
+	            fi
 	            exit 1
 	            ;;
 	    esac
